requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
cryptography==43.0.0
Django==5.1
django-crispy-forms==2.3
django-paypal==2.1
idna==3.8
pycparser==2.22
pyOpenSSL==24.2.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
